Updates on DirectoryOperations class

+++ DESCRIPTOR.FILE.READER/Descriptor_File_Data_Collector/Descriptor_File_Data_Collector.cpp
@@ -122,7 +122,7 @@ void Descriptor_File_Data_Collector::Clear_Dynamic_Memory(){
 
 void Descriptor_File_Data_Collector::Receive_Descriptor_File_Directory(const char * DescriptorFileDirectory){
 
-     int String_Size = strlen(DescriptorFileDirectory);
+     size_t String_Size = strlen(DescriptorFileDirectory);
 
      this->Memory_Delete_Condition = false;
 
@@ -135,7 +135,7 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Directory(const cha
 
 void Descriptor_File_Data_Collector::Receive_Descriptor_File_Directory(char * DescriptorFileDirectory){
 
-     int String_Size = strlen(DescriptorFileDirectory);
+     size_t String_Size = strlen(DescriptorFileDirectory);
 
      this->Memory_Delete_Condition = false;
 
@@ -148,7 +148,7 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Directory(char * De
 
 void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(const char * DescriptorFileName){
 
-     int String_Size = strlen(DescriptorFileName);
+     size_t String_Size = strlen(DescriptorFileName);
 
      this->Memory_Delete_Condition = false;
 
@@ -170,7 +170,7 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(const char * D
 
 void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * DescriptorFileName){
 
-     int String_Size = strlen(DescriptorFileName);
+     size_t String_Size = strlen(DescriptorFileName);
 
      this->Memory_Delete_Condition = false;
 
@@ -618,33 +618,33 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * Descrip
 
       char end_brace [] = "]";
 
-      int start_word_size = strlen(start_word);
+      size_t start_word_size = strlen(start_word);
 
-      int end_brace_size = strlen(end_brace);
+      size_t end_brace_size = strlen(end_brace);
 
-      int Start_Point_Size = strlen(Start_Point);
+      size_t Start_Point_Size = strlen(Start_Point);
 
-      int search_string_size = start_word_size + end_brace_size + Start_Point_Size;
+      size_t search_string_size = start_word_size + end_brace_size + Start_Point_Size;
 
       char * search_string = new char [10*search_string_size];
 
       int index_counter = 0;
 
-      for(int i=0;i<start_word_size;i++){
+      for(size_t i=0;i<start_word_size;i++){
 
           search_string[index_counter] = start_word[i];
 
           index_counter++;
       }
 
-      for(int i=0;i<Start_Point_Size;i++){
+      for(size_t i=0;i<Start_Point_Size;i++){
 
           search_string[index_counter] = Start_Point[i];
 
           index_counter++;
       }
 
-      for(int i=0;i<end_brace_size;i++){
+      for(size_t i=0;i<end_brace_size;i++){
 
           search_string[index_counter] = end_brace[i];
 
@@ -688,9 +688,9 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * Descrip
 
           char * String_Line = this->StringOperations.GetStringBuffer();
 
-          int String_Size = strlen(String_Line);
+          size_t String_Size = strlen(String_Line);
 
-          for(int i=0;i<String_Size;i++){
+          for(size_t i=0;i<String_Size;i++){
 
               if(String_Line[i] == '{'){
 
@@ -730,7 +730,7 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * Descrip
          this->ConstString = nullptr;
       }
 
-      int String_Size = strlen(ConstString);
+      size_t String_Size = strlen(ConstString);
 
       this->ConstString = new char [10*String_Size];
 
@@ -870,11 +870,11 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * Descrip
 
          this->StringOperations.ReadFileLine(i);
 
-         int String_Size = strlen(this->StringOperations.GetStringBuffer());
+         size_t String_Size = strlen(this->StringOperations.GetStringBuffer());
 
          bool is_it_a_brace = false;
 
-         for(int k=0;k<String_Size;k++){
+         for(size_t k=0;k<String_Size;k++){
 
              if(((this->StringOperations.GetStringBuffer()[k] == '{') || (this->StringOperations.GetStringBuffer()[k] == '}'))){
 
@@ -886,9 +886,9 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * Descrip
 
          while((this->StringOperations.GetStringBuffer()[0] == '\t') || (this->StringOperations.GetStringBuffer()[0] == '\b') || (this->StringOperations.GetStringBuffer()[0] == ' ')){
 
-               int String_Size = strlen(this->StringOperations.GetStringBuffer());
+               size_t String_Size = strlen(this->StringOperations.GetStringBuffer());
 
-               for(int k=0;k<String_Size;k++){
+               for(size_t k=0;k<String_Size;k++){
 
                    this->StringOperations.GetStringBuffer()[k] = this->StringOperations.GetStringBuffer()[k+1];
                }
@@ -915,18 +915,18 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * Descrip
       }
  }
 
- void Descriptor_File_Data_Collector::Place_String(char ** Pointer, char * String, int String_Size){
+ void Descriptor_File_Data_Collector::Place_String(char ** Pointer, char * String, size_t String_Size){
 
-      int Start_Point = 0;
+      size_t Start_Point = 0;
 
       while(((String[Start_Point]== ' ') || (String[Start_Point]== '\t'))){
 
             Start_Point++;
       }
 
-      int index_counter = 0;
+      size_t index_counter = 0;
 
-      for(int i=Start_Point;i<String_Size;i++){
+      for(size_t i=Start_Point;i<String_Size;i++){
 
           (*Pointer)[index_counter] = String[i];
 
@@ -936,18 +936,18 @@ void Descriptor_File_Data_Collector::Receive_Descriptor_File_Name(char * Descrip
       (*Pointer)[index_counter] = '\0';
  }
 
- void Descriptor_File_Data_Collector::Place_String(char ** Pointer, const char * String, int String_Size){
+ void Descriptor_File_Data_Collector::Place_String(char ** Pointer, const char * String, size_t String_Size){
 
-      int Start_Point = 0;
+      size_t Start_Point = 0;
 
       while(((String[Start_Point]== ' ') || (String[Start_Point]== '\t'))){
 
               Start_Point++;
       }
 
-      int index_counter = 0;
+      size_t index_counter = 0;
 
-      for(int i=Start_Point;i<String_Size;i++){
+      for(size_t i=Start_Point;i<String_Size;i++){
 
          (*Pointer)[index_counter] = String[i];
